+++ src/main/java/kitchenpos/eatin/application/OrderService.java
@@ -1,7 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.eatin.application;
 
-import kitchenpos.domain.*;
+import kitchenpos.eatin.domain.*;
 import kitchenpos.infra.KitchenridersClient;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
