+++ src/main/java/kitchenpos/eatin/domain/Order.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatin.domain;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
