+++ src/main/java/kitchenpos/eatin/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatin.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
