+++ src/main/java/kitchenpos/eatin/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatin.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
