+++ src/main/java/kitchenpos/eatin/infra/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.eatin.infra;
 
+import kitchenpos.eatin.domain.Order;
+import kitchenpos.eatin.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
