+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
