+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,15 +1,14 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.math.BigDecimal;
 import java.util.UUID;
 
-@Table(name = "product")
+@Table(name = "menu_group")
 @Entity
-public class Product {
+public class MenuGroup {
     @Column(name = "id", columnDefinition = "binary(16)")
     @Id
     private UUID id;
@@ -17,10 +16,7 @@ public class Product {
     @Column(name = "name", nullable = false)
     private String name;
 
-    @Column(name = "price", nullable = false)
-    private BigDecimal price;
-
-    public Product() {
+    public MenuGroup() {
     }
 
     public UUID getId() {
@@ -37,13 +33,5 @@ public class Product {
 
     public void setName(final String name) {
         this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
     }
 }
